+++ atividades/Trabalho 04/Questão 01/Multiplication.java
@@ -6,7 +6,7 @@ class Multiplication implements Expression {
         this.right = r;
     }
 
-    public float calculate(){
+    public double calculate(){
         return this.left.calculate() * this.right.calculate();
     }
 }
